move existing ds methods to it's own file
